Revert "0802调试前"
This reverts commit 593df8d15dc1a805dfb30ccfdc64e5e54e883487.

+++ FanucCnc/Model/ParaDieChangeInfo.cs
@@ -25,6 +25,6 @@ namespace FanucCnc.Model
 
         public double DieWeight { get; set; }
 
-        public double LoaderSafePosition { get; set; }
+        public int LoaderSafePosition { get; set; }
     }
 }
